+++ src/test/java/nextstep/subway/acceptance/LineAcceptanceTest.java
@@ -1,36 +1,36 @@
 package nextstep.subway.acceptance;
 
-import static nextstep.subway.fixture.given.LineRequestFixture.distance;
-import static nextstep.subway.fixture.given.LineRequestFixture.green;
-import static nextstep.subway.fixture.given.LineRequestFixture.red;
-import static nextstep.subway.fixture.given.LineRequestFixture.다른분당선;
-import static nextstep.subway.fixture.given.LineRequestFixture.분당선;
-import static nextstep.subway.fixture.given.LineRequestFixture.새로운지하철역_id;
-import static nextstep.subway.fixture.given.LineRequestFixture.신분당선;
-import static nextstep.subway.fixture.given.LineRequestFixture.지하철역_id;
-import static nextstep.subway.fixture.given.StationModifyRequestFixture.노선수정요청_데이터_생성;
-import static nextstep.subway.fixture.given.StationModifyRequestFixture.또다른지하철역이름;
-import static nextstep.subway.fixture.given.StationModifyRequestFixture.새로운지하철역이름;
-import static nextstep.subway.fixture.given.StationModifyRequestFixture.지하철역이름;
-import static nextstep.subway.fixture.then.ApiStatusFixture.API_생성_응답코드_검사;
-import static nextstep.subway.fixture.then.ApiStatusFixture.API_요청성공_응답코드_검사;
-import static nextstep.subway.fixture.then.ApiStatusFixture.API_잘못된요청_응답코드_검사;
-import static nextstep.subway.fixture.then.LineThenFixture.노선_응답값_id_검사;
-import static nextstep.subway.fixture.then.LineThenFixture.노선_응답값_노선색_검사;
-import static nextstep.subway.fixture.then.LineThenFixture.노선_응답값_노선이름_검사;
-import static nextstep.subway.fixture.then.LineThenFixture.노선_응답값_포함된_지하철역_크기_검사;
-import static nextstep.subway.fixture.then.LineThenFixture.노선_응답값_포함된지하철역_id_검사;
-import static nextstep.subway.fixture.then.LineThenFixture.노선목록_노선이름_포함_검사;
-import static nextstep.subway.fixture.then.LineThenFixture.노선목록_조회시_생성한노선_id_포함_검사;
-import static nextstep.subway.fixture.then.LineThenFixture.노선목록_크기_검사;
-import static nextstep.subway.fixture.then.LineThenFixture.없는노선조회시_에러표출_검사;
-import static nextstep.subway.fixture.when.LineApiFixture.지하철역_노선_단건_조회;
-import static nextstep.subway.fixture.when.LineApiFixture.지하철역_노선_등록_요청;
-import static nextstep.subway.fixture.when.LineApiFixture.지하철역_노선_등록_요청_후_id_추출;
-import static nextstep.subway.fixture.when.LineApiFixture.지하철역_노선_목록_조회_요청;
-import static nextstep.subway.fixture.when.LineApiFixture.지하철역_노선_삭제;
-import static nextstep.subway.fixture.when.LineApiFixture.지하철역_노선_수정;
-import static nextstep.subway.fixture.when.StationApiFixture.지하철역_생성_요청;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.distance;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.green;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.red;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.다른분당선;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.분당선;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.새로운지하철역_id;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.신분당선;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.지하철역_id;
+import static nextstep.subway.fixture.acceptance.given.StationModifyRequestFixture.노선수정요청_데이터_생성;
+import static nextstep.subway.fixture.acceptance.given.StationModifyRequestFixture.또다른지하철역이름;
+import static nextstep.subway.fixture.acceptance.given.StationModifyRequestFixture.새로운지하철역이름;
+import static nextstep.subway.fixture.acceptance.given.StationModifyRequestFixture.지하철역이름;
+import static nextstep.subway.fixture.acceptance.then.ApiStatusFixture.API_생성_응답코드_검사;
+import static nextstep.subway.fixture.acceptance.then.ApiStatusFixture.API_요청성공_응답코드_검사;
+import static nextstep.subway.fixture.acceptance.then.ApiStatusFixture.API_잘못된요청_응답코드_검사;
+import static nextstep.subway.fixture.acceptance.then.LineThenFixture.노선_응답값_id_검사;
+import static nextstep.subway.fixture.acceptance.then.LineThenFixture.노선_응답값_노선색_검사;
+import static nextstep.subway.fixture.acceptance.then.LineThenFixture.노선_응답값_노선이름_검사;
+import static nextstep.subway.fixture.acceptance.then.LineThenFixture.노선_응답값_포함된_지하철역_크기_검사;
+import static nextstep.subway.fixture.acceptance.then.LineThenFixture.노선_응답값_포함된지하철역_id_검사;
+import static nextstep.subway.fixture.acceptance.then.LineThenFixture.노선목록_노선이름_포함_검사;
+import static nextstep.subway.fixture.acceptance.then.LineThenFixture.노선목록_조회시_생성한노선_id_포함_검사;
+import static nextstep.subway.fixture.acceptance.then.LineThenFixture.노선목록_크기_검사;
+import static nextstep.subway.fixture.acceptance.then.LineThenFixture.없는노선조회시_에러표출_검사;
+import static nextstep.subway.fixture.acceptance.when.LineApiFixture.지하철역_노선_단건_조회;
+import static nextstep.subway.fixture.acceptance.when.LineApiFixture.지하철역_노선_등록_요청;
+import static nextstep.subway.fixture.acceptance.when.LineApiFixture.지하철역_노선_등록_요청_후_id_추출;
+import static nextstep.subway.fixture.acceptance.when.LineApiFixture.지하철역_노선_목록_조회_요청;
+import static nextstep.subway.fixture.acceptance.when.LineApiFixture.지하철역_노선_삭제;
+import static nextstep.subway.fixture.acceptance.when.LineApiFixture.지하철역_노선_수정;
+import static nextstep.subway.fixture.acceptance.when.StationApiFixture.지하철역_생성_요청;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import io.restassured.response.ExtractableResponse;
