+++ src/test/java/nextstep/subway/acceptance/SectionAcceptanceTest.java
@@ -1,27 +1,27 @@
 package nextstep.subway.acceptance;
 
-import static nextstep.subway.fixture.given.LineRequestFixture.distance;
-import static nextstep.subway.fixture.given.LineRequestFixture.red;
-import static nextstep.subway.fixture.given.LineRequestFixture.신분당선;
-import static nextstep.subway.fixture.given.SectionRequestFixture.구간거리;
-import static nextstep.subway.fixture.given.StationModifyRequestFixture.또다른지하철역이름;
-import static nextstep.subway.fixture.given.StationModifyRequestFixture.새로운지하철역이름;
-import static nextstep.subway.fixture.given.StationModifyRequestFixture.지하철역이름;
-import static nextstep.subway.fixture.then.ApiStatusFixture.API_삭제_응답코드_검사;
-import static nextstep.subway.fixture.then.ApiStatusFixture.API_생성_응답코드_검사;
-import static nextstep.subway.fixture.then.ApiStatusFixture.API_잘못된요청_응답코드_검사;
-import static nextstep.subway.fixture.then.SectionThenFixture.노선구간추가시_구간거리가_기존거리보다_같거나_길다면_에러;
-import static nextstep.subway.fixture.then.SectionThenFixture.노선구간추가시_모든역이_노선에_이미_존재할때_오류_검사;
-import static nextstep.subway.fixture.then.SectionThenFixture.노선구간추가시_모든역이_노선에_존재하지않을때_오류_검사;
-import static nextstep.subway.fixture.then.SectionThenFixture.삭제할_노선_구간_1개인경우_에러;
-import static nextstep.subway.fixture.then.SectionThenFixture.삭제할_노선_구간이_하행종점역이_아닐경우_에러;
-import static nextstep.subway.fixture.then.SectionThenFixture.지하철_노선_조회시_구간_id_순서_검사;
-import static nextstep.subway.fixture.then.SectionThenFixture.지하철_노선_조회시_구간포함_확인;
-import static nextstep.subway.fixture.then.SectionThenFixture.지하철_노선_조회시_해당구간_불포함_확인;
-import static nextstep.subway.fixture.when.LineApiFixture.지하철역_노선_등록_요청_후_id_추출;
-import static nextstep.subway.fixture.when.SectionApiFixture.지하철_노선_구간_삭제;
-import static nextstep.subway.fixture.when.SectionApiFixture.지하철_노선_구간_추가_등록;
-import static nextstep.subway.fixture.when.StationApiFixture.지하철역_생성_요청;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.distance;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.red;
+import static nextstep.subway.fixture.acceptance.given.LineRequestFixture.신분당선;
+import static nextstep.subway.fixture.acceptance.given.SectionRequestFixture.구간거리;
+import static nextstep.subway.fixture.acceptance.given.StationModifyRequestFixture.또다른지하철역이름;
+import static nextstep.subway.fixture.acceptance.given.StationModifyRequestFixture.새로운지하철역이름;
+import static nextstep.subway.fixture.acceptance.given.StationModifyRequestFixture.지하철역이름;
+import static nextstep.subway.fixture.acceptance.then.ApiStatusFixture.API_삭제_응답코드_검사;
+import static nextstep.subway.fixture.acceptance.then.ApiStatusFixture.API_생성_응답코드_검사;
+import static nextstep.subway.fixture.acceptance.then.ApiStatusFixture.API_잘못된요청_응답코드_검사;
+import static nextstep.subway.fixture.acceptance.then.SectionThenFixture.노선구간추가시_구간거리가_기존거리보다_같거나_길다면_에러;
+import static nextstep.subway.fixture.acceptance.then.SectionThenFixture.노선구간추가시_모든역이_노선에_이미_존재할때_오류_검사;
+import static nextstep.subway.fixture.acceptance.then.SectionThenFixture.노선구간추가시_모든역이_노선에_존재하지않을때_오류_검사;
+import static nextstep.subway.fixture.acceptance.then.SectionThenFixture.삭제할_노선_구간_1개인경우_에러;
+import static nextstep.subway.fixture.acceptance.then.SectionThenFixture.삭제할_노선_구간이_하행종점역이_아닐경우_에러;
+import static nextstep.subway.fixture.acceptance.then.SectionThenFixture.지하철_노선_조회시_구간_id_순서_검사;
+import static nextstep.subway.fixture.acceptance.then.SectionThenFixture.지하철_노선_조회시_구간포함_확인;
+import static nextstep.subway.fixture.acceptance.then.SectionThenFixture.지하철_노선_조회시_해당구간_불포함_확인;
+import static nextstep.subway.fixture.acceptance.when.LineApiFixture.지하철역_노선_등록_요청_후_id_추출;
+import static nextstep.subway.fixture.acceptance.when.SectionApiFixture.지하철_노선_구간_삭제;
+import static nextstep.subway.fixture.acceptance.when.SectionApiFixture.지하철_노선_구간_추가_등록;
+import static nextstep.subway.fixture.acceptance.when.StationApiFixture.지하철역_생성_요청;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
