+++ src/test/java/nextstep/subway/fixture/acceptance/then/SectionThenFixture.java
@@ -1,6 +1,6 @@
-package nextstep.subway.fixture.then;
+package nextstep.subway.fixture.acceptance.then;
 
-import static nextstep.subway.fixture.when.LineApiFixture.지하철역_노선_단건_조회;
+import static nextstep.subway.fixture.acceptance.when.LineApiFixture.지하철역_노선_단건_조회;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.restassured.response.ExtractableResponse;
